`ReadModel`, make `where` and `order` just use `merge`
This changes the read model `where` and `order` methods to just
be alias of `merge`. This means that a read models `where` and
`order` cannot take arguments for ActiveRecord's matching relation
methods. Instead, these 2 querying methods have to be passed
relations so they can be merged onto the main query's relation.

In our usage of read models, when a where or order was needed, a
defined scope was just merged on. Thus, it's more readable to
be able to use `where` and `order` instead of just `merge`. This
also promotes preferred conventions of using defined scopes or at
least anchoring a where or order on a record (a table), which
makes a read model easier to comprehend.

Closes #126