Add `find` method to `ReadModel`
This adds a `find` method to `ReadModel`. This will allow querying
for a specific record in the collection. It will use
ActiveRecord's `find` method on the relation. The relation itself
is only built from a subset of the read model's expressions. It
will not use any `where`, `order`, `limit` or `offset` expressions.
This is because `order` doesn't matter when selecting a specific
record and `where`, `limit` and `offset` can remove the record from
the results.

Closes #142