Bug fix, `assert_association_saved` use actual record's class
This updates the `assert_association_saved` test helper to use the
model's record's class instead of using configured `record_class`.
This was a problem with model's configured with "real" ActiveRecord
records but whose real records were an instance of `FakeRecord`.
The test helper would expect the wrong value for a polymorphic
belongs to's foreign type column. This fixes that by not using
what is configured but instead using what the actual record's
class is.