Remove "caching" of `Factory` date, datetime and time objects
This changes `MR::Factory` to no longer "cache" date, datetime
and time objects. This was previously done as a speed enhancement
and to allow "always working from a known date/time". The speed
enhancement was nice, but we've decided that it's more consistent
if it always tries to generate random data. The "always working
from a known date/time" has proved to not be an issue.