Have `RecordStack` reset fake records `save_called` flag
This updates `RecordStack` to reset a fake records `save_called`
flag after it creates it. This ensures that a stack doesn't
interfere with testing. If this isn't done, then anytime a stack
creates a fake model, it's `save_called` flag is `true`. This
can cause false positive tests when using `assert_model_saved`.
This is only done for fake records and is done for every fake
record that the stack creates.