read model: have `find` return instances of the read model
This updates the newly added read model `find` method to, in addition
to running find on the basic `from` record, use that record data
result to build and return an instance of the read model as its
result.

The previous implementation just blindly returned the result of
the from record's `find` method.  While this result would have all
of the custom select data available, it would be lacking any mixed
in methods that were on the read model.  Plus, it didn't follow the
MR pattern of "get the raw record and create a model/read-model from
it".

This updates the tests for the find to be a little more robust.  They
had no "kind of" test to ensure a read model instance was being
returned.  I added this and the tests still passed.  This was due
to how the tests were stubbing results.  I changed it so that the
stubbed returned result was a fake record instead of a read model
instance.  This is more accurate for what is happening in the code.