Support grouped counts with `MR::PagedQuery#total_count`
This updates the `MR::PagedQuery` to work with grouped counts
returned by a relation. The query will sum the values of the
result hash to get the same value as an ungrouped query. This
causes issues with performing proper paging calculations if this
value returns a hash and not an integer.

/cc @kellyredding