Update record stack, preload preset associations into lookup
This updates record stacks to preload preset associations into
its dependency lookup. This has the benefit of making sure that
a stack uses an already created association when possible.
Previously, depending on the order it walked associations, it
could build a new type of record even though a preset association
already had an instance of that type of record. This ended up
being an annoyance with the behavior and doesn't fit well with
its theme of only creating the minimal information needed. This
shouldn't break any existing behavior, just allows stacks to
be created with less information.

Closes #112