Add `column_for_attribute` to `FakeRecord`
This adds the ActiveRecord method `column_for_attribute` to
`FakeRecord`. This is useful for getting the column definition for
an attribute from an instance of an ActiveRecord model. This will
be used by record stacks to see if an association's columns are
required or not (whether they allow null values). This method was
added so both real and fake records will work with record stacks.