Add `saved_instance` factory method
This adds a `saved_instance` factory method for model and record
factories. This will automatically save the model/record instead
of having to manually do this. More importantly, this also handles
resetting the `save_called` flag for fake records. Because stacks
are already resetting this flag, this should make it unecessary to
call the test helper `model_reset_save_called`.