Rename `ReadModel` private methods to avoid collisions
This renames the `ReadModel` data private methods to avoid
collisions. Originally these were named with the generic names
`data` and `set_data`. This is unfortunate in the case something
using the `ReadModel` mixin wants to have a `data` method. This
fixes the issue by naming the method something specific as they
are only intended for internal use.