remove the deprecated `models` method on query objects
We originally had `models` in place as the method to fetch the
results (which were "models" or "read models").  In practice though
it seemed awkward to call `models` on a query b/c "models" isn't
typically the language you use to talk about query results - "results"
is.

So we aliased the `results` method and began switching to using
that instead.  Now is the time to remove this legacy method and
fully switch to using `results` fully.

This removes the alias and renames the method to `results`.  It then
goes through the tests and updates their languange/descs to match.
In a few cases I cleaned up the descs to make them more consistent.