Fix benchmarks from failing because of DB constraints
This fixes the benchmarks from failing due to DB constraints.
The DB constraints were added to test how record stacks handled
required and optional associations. Before this, all the
associations were optional and this is the assumption the
benchmark scripts were written with. Since this assumption is no
longer true, they needed to be updated. This updates to create
any models with the required associations and adds optional
associations that the benchmark scripts can freely set and unset
as needed.

This also re-runs all the benchmarks to create a new reference
point to compare future changes against.