Remove Travis CI configuration
This removes Travis CI configuration and I've turned it off through
their website. I wanted to try using it out and see how it works.
It doesn't add much value to our current work flow for testing
different ruby versions so I'm removing it to avoid the overhead
of keeping it updated and working.

/cc @kellyredding.