Define `type` method on `SubQueryExpression`
This adds a `type` method on `SubQueryExpression` to make it a
compatible query expression. This was causing a warning whenever
a read model's `find` method was used and the read model used
subqueries. This is because it would call the ruby provided `type`
method which throws a warning because it has been deprecated. All
query expressions are expected to have a `type` method but this
was missed when adding the subquery logic.