Update Record Factory to use `sql_type` if `type` is missing
This updates the Record Factory to use the ActiveRecord Column's
`sql_type` if `type` is missing/nil.  This makes MR more flexible in
which columns can be used.  If you are using a field in the database
that isn't support by ActiveRecord (i.e. `json` or `jsonb` for
ActiveRecord earlier than version 4), you can define a factory that
matches the `sql_type` and still use the factory.

This also updates the factory to avoid setting any fields that are not
supported.  This is the least suprising behavior as it does not force the
end user to define every columne type that is used, just those that are
important enough to generate data for.